chore: align (and adjust) nullability
