+++ backend-cm/src/investigation/investigation_contact_method/investigation_contact_method_inputs.graphql
@@ -1,13 +1,13 @@
 input CreateInvestigationContactMethodInput {
   contactMethodGuid: String
   typeCode: String!
-  value: String
-  isPrimary: Boolean
+  value: String!
+  isPrimary: Boolean!
 }
 
 input UpdateInvestigationContactMethodInput {
-  contactMethodGuid: String
+  contactMethodGuid: String!
   typeCode: String!
-  value: String
-  isPrimary: Boolean
+  value: String!
+  isPrimary: Boolean!
 }
