+++ backend-cm/src/investigation/investigation_contact_method/investigation_contact_method_types.graphql
@@ -1,6 +1,6 @@
 type InvestigationContactMethod {
   contactMethodGuid: String!
   typeCode: String!
-  value: String
+  value: String!
   isPrimary: Boolean!
 }
