+++ backend-cm/src/shared/contact_method/contact_method_types.graphql
@@ -1,7 +1,7 @@
 type ContactMethod {
-  contactMethodGuid: String
-  typeCode: String
-  typeDescription: String
-  value: String
-  isPrimary: Boolean
+  contactMethodGuid: String!
+  typeCode: String!
+  typeDescription: String!
+  value: String!
+  isPrimary: Boolean!
 }
