+++ frontend/src/app/components/containers/parties/view/party-view.tsx
@@ -5,12 +5,11 @@ import { PartyTabs } from "./party-tabs";
 import { PartyHistoryTab } from "./party-history-tab";
 import { useGraphQLQuery } from "@graphql/hooks";
 import { gql } from "graphql-request";
-import { BusinessIdentifier, Party } from "@/generated/graphql";
+import { Party } from "@/generated/graphql";
 import { Button } from "react-bootstrap";
 import { PartyTypes } from "@/app/constants/party-types";
 import { PartyComplianceHistory } from "@/app/components/containers/parties/view/compliance-history/party-compliance-history";
-import PartyDetail from "@/app/components/containers/parties/view/party-detail";
-import { DetailSection } from "@/app/components/containers/investigations/details/investigation-parties/investigation-party-view";
+import PartyDetail, { DetailSection } from "@/app/components/containers/parties/view/party-detail";
 import { PartyAttachments } from "@/app/components/containers/parties/attachments/party-attachments";
 import AttachmentEnum from "@/app/constants/attachment-enum";
 import { isYoungPerson } from "@/app/common/methods";
